moves components into their own files
