+++ src/components/footer.js
@@ -1,0 +1,12 @@
+import React from 'react';
+
+const Footer = () => {
+  return (
+    <ul>
+      <li>&copy; Copyright Team Banana/Guava 2017</li>
+      <li>About</li>
+    </ul>
+  )
+}
+
+export default Footer;
