+++ src/components/home/RecruiterIntro.js
@@ -1,0 +1,14 @@
+import React from 'react';
+
+const RecruiterIntro = () => {
+  return (
+    <div>
+      <p>
+        Recruiter instructions here
+      </p>
+      <button>button to sign up</button>
+    </div>
+  )
+}
+
+export default RecruiterIntro;
