+++ src/components/home/SeekerIntro.js
@@ -1,0 +1,14 @@
+import React from 'react';
+
+const SeekerIntro = () => {
+  return (
+    <div>
+      <p>
+        Seeker instructions here
+      </p>
+      <button>button to sign up</button>
+    </div>
+  )
+}
+
+export default SeekerIntro;
