+++ src/components/home/Welcome.js
@@ -1,0 +1,11 @@
+import React from 'react';
+
+const Welcome = () => {
+  return (
+    <div>
+      <p>Welcome message from JobOrNot</p>
+    </div>
+  )
+}
+
+export default Welcome;
