+++ src/components/navbar/SignedInNav.js
@@ -1,0 +1,17 @@
+import React from 'react';
+
+const SignedInNav = (props) => {
+  return (
+    <nav>
+      <ul>
+        <li><img src="http://placehold.it/50x50" alt='logo' /></li>
+        <li>JobOrNot</li>
+        <li>How It Works</li>
+        <li>Profile</li>
+        <li>{props.button}</li>
+      </ul>
+    </nav>
+  )
+}
+
+export default SignedInNav;
